+++ graphql.go
@@ -18,7 +18,7 @@ type VariationQL struct {
 func GetExperimentsGraphQL(siteCode string) string {
 	return `{
 		"operationName": "getExperiments",
-		"query": "query getExperiments($first: Int, $after: String, $filter: FilteringExpression, $sort: [SortingParameter!]) { experiments(first: $first, after: $after, filter: $filter, sort: $sort) { edges { node { id name type site { id code } status variations { id customJson } deviations { variationId value } respoolTime {variationId value } segment { id name conditionsData { firstLevelOrOperators firstLevel { orOperators conditions { targetingType isInclude ... on CustomDataTargetingCondition { customDataIndex value valueMatchType } } } } } __typename } __typename } pageInfo { endCursor hasNextPage __typename } totalCount __typename } }",
+		"query": "query getExperiments($first: Int, $after: String, $filter: FilteringExpression, $sort: [SortingParameter!]) { experiments(first: $first, after: $after, filter: $filter, sort: $sort) { edges { node { id name type site { id code isKameleoonEnabled } status variations { id customJson } deviations { variationId value } respoolTime {variationId value } segment { id name conditionsData { firstLevelOrOperators firstLevel { orOperators conditions { targetingType isInclude ... on CustomDataTargetingCondition { customDataIndex value valueMatchType } } } } } __typename } __typename } pageInfo { endCursor hasNextPage __typename } totalCount __typename } }",
 		"variables": {
 			"filter": {
 				"and": [{
@@ -111,10 +111,10 @@ func (expNodeQL *NodeExperimentGraphQL) Transform() types.Experiment {
 	return exp
 }
 
-func GetFeatureFlagsGraphQL(siteCode string) string {
+func GetFeatureFlagsGraphQL(siteCode string, environment string) string {
 	return `{
 		"operationName": "getFeatureFlags",
-		"query": "query getFeatureFlags($first: Int, $after: String, $filter: FilteringExpression, $sort: [SortingParameter!]) { featureFlags(first: $first, after: $after, filter: $filter, sort: $sort) { edges { node { id name site { id code } bypassDeviation status variations { id customJson } respoolTime { variationId value } expositionRate identificationKey featureFlagSdkLanguageType featureStatus schedules { dateStart dateEnd } segment { id name conditionsData { firstLevelOrOperators firstLevel { orOperators conditions { targetingType isInclude ... on CustomDataTargetingCondition { customDataIndex value valueMatchType } } } } } __typename } __typename } pageInfo { endCursor hasNextPage __typename } totalCount __typename } }",
+		"query": "query getFeatureFlags($first: Int, $after: String, $filter: FilteringExpression, $sort: [SortingParameter!]) { featureFlags(first: $first, after: $after, filter: $filter, sort: $sort) { edges { node { id name site { id code isKameleoonEnabled } bypassDeviation status variations { id customJson } respoolTime { variationId value } expositionRate identificationKey featureFlagSdkLanguageType featureStatus schedules { dateStart dateEnd } segment { id name conditionsData { firstLevelOrOperators firstLevel { orOperators conditions { targetingType isInclude ... on CustomDataTargetingCondition { customDataIndex value valueMatchType } } } } } __typename } __typename } pageInfo { endCursor hasNextPage __typename } totalCount __typename } }",
 		"variables": {
 			"filter": {
 				"and": [{
@@ -123,11 +123,19 @@ func GetFeatureFlagsGraphQL(siteCode string) string {
 						"operator": "IN",
 						"parameters": ["ACTIVATED", "SCHEDULED", "DEACTIVATED"]
 					}
-				}, {
+				}, 
+				{
 					"condition": {
 						"field": "siteCode",
 						"operator": "IN",
 						"parameters": ["` + siteCode + `"]
+					}
+				},
+				{
+					"condition": {
+						"field": "environment.key",
+						"operator": "IN",
+						"parameters": ["` + environment + `"]
 					}
 				}]
 			},